新增：修改功能和整合代码rev, vcc, temp

+++ keil_projects/proj_template/proj_template/my_app/temperature.h
@@ -38,7 +38,7 @@ extern TemperReadCfg_t g_temperReadCfg;
 int8 temper_getTemperValue(uint16 cnt);
 uint16 temper_getTemperCnt(void);
 void temper_sampleTemperTimerCb(void);
-int8 temper_sampleTemper(void);
+float temper_sampleTemper(void);
 void temper_resetInit(void);
 
 #endif //__TEMPERATURE_H_
